requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
mysql-connector-python==8.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
sendgrid==6.11.0
fpdf2==2.7.8
jinja2==3.1.4
email-validator==2.1.1
pytest==8.2.0
httpx==0.27.0
